Split the events module up into muiltiple files
